RDKOSS-490: Move OSS_LAYER_VERSION to reference layer
Reason for the change - The meta-oss-reference-release repository has been excluded from the manifest to support the new OSS consumption model. Consequently, the OSS_LAYER_VERSION parameter previously defined in that repository is no longer available to the build.
To address this, the OSS_LAYER_VERSION parameter has been moved to the reference layer to ensure continued availability during the build process.